[Refactor/#116]: ProfileImage 분리
