+++ src/features/Feed/components/CommentList/ChildCommentItem.tsx
@@ -4,7 +4,7 @@ import React from 'react';
 import { useSearchParams } from 'next/navigation';
 import { useProfile } from '@/hooks/useProfile';
 import useCommentStore from '@/store/commentStore';
-import ProfileHeader from '@/components/Header/ProfileHeader';
+import ProfileHeader from '@/components/Header/ProfileHeader/ProfileHeader';
 import LikeButton from '@/components/Button/LikeButton';
 import { motion } from 'framer-motion';
 import { formatDate } from '@/utils/date';
