+++ src/features/Feed/components/CommentList/CommentItem.tsx
@@ -8,7 +8,7 @@ import useCommentStore from '@/store/commentStore';
 import LikeButton from '@/components/Button/LikeButton';
 import { useProfile } from '@/hooks/useProfile';
 import { formatDate } from '@/utils/date';
-import ProfileHeader from '@/components/Header/ProfileHeader';
+import ProfileHeader from '@/components/Header/ProfileHeader/ProfileHeader';
 import ChildCommentItem from './ChildCommentItem';
 import { useChildComments } from '../../hooks/child/useChildComments';
 import { Comments } from '../../types/comment';
